update user role

+++ application/controllers/Process.php
@@ -4,19 +4,20 @@ defined('BASEPATH') OR exit('No direct script access allowed');
 class Process extends CI_Controller {
     function __construct() {
         parent::__construct();
-        $udata = $this->session->userdata();
-        if((!isset($udata['username'])) && (!isset($udata['id'])))die;
+        $this->load->model('auths');
     }
 
     private function check(){
     }
 
     public function delete($table, $id) {
+        if(!$this->auths->admin()) die;
         $this->db->where('id', $id)->delete($table);
         echo "done";
     }
 
     public function insert($table) {
+        if(!$this->auths->admin()) die;
         $data = $this->input->post();
         if($table == "users"){
             if(!isset($data['username']))die;
@@ -41,6 +42,25 @@ class Process extends CI_Controller {
     
     public function edit($table, $id) {
         $data = $this->input->post();
+        if($table == "users"){
+            if($id != $_SESSION['id']){
+                if(!$this->auths->admin()) die;
+            }
+
+            if(!isset($data['username']))die;
+
+            $us = $this->db->where([
+                'username' => $data['username'],
+                'id !='    => $id 
+                ])->get('users')->result();
+
+                if(isset($us[0])){
+                echo 'OOPS, Username has been used by another user!';
+                die;
+            }
+        } else {
+            if(!$this->auths->admin()) die;
+        }
         foreach($data as $key => $val ){
             if($key == "password"){
                 if($val == ""){
